Update README.md
This change is to make sure the install.sh file completes running. This repo must be cloned as marsha rather than MARSHA so that the mv ~/marsha ~/catkin_ws/src command recognizes the cloned repo and moves it. I re-cloned using the git clone command here with this change and the install.sh file ran all the way through.